更新作者页面样式

+++ QianShi.Music/Common/Models/Quality.cs
@@ -6,7 +6,7 @@
     public record Quality
     {
         public int Br { get; set; }
-        public int Fid { get; set; }
+        public double Fid { get; set; }
         public long Size { get; set; }
         public int Sr { get; set; }
         public double Vd { get; set; }
